+++ index.js
@@ -2,7 +2,18 @@ require('module-alias/register');
 require('dotenv').config()
 
 const Discord = require("discord.js");
-const client = new Discord.Client();
+const {
+  Client,
+  Collection,
+  Intents
+} = require('discord.js');
+
+// !This is the new way to create the client.
+// I just pass in majority of the Intents incase I do use them along the way
+// I believe its preferred to pass in the intents you are using then all.
+const client = new Client({
+  intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MEMBERS, Intents.FLAGS.GUILD_BANS, Intents.FLAGS.GUILD_MESSAGE_REACTIONS, Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.GUILD_MESSAGE_REACTIONS, Intents.FLAGS.DIRECT_MESSAGES, Intents.FLAGS.DIRECT_MESSAGE_REACTIONS]
+});
 //this should be initialized strictly after client is define
 client.config = require("@root/config.js");
 
@@ -13,8 +24,8 @@ const loadModules = require("@handlers/loadModules.js");
 const functions = require("@modules/functions.js");
 functions(client);
 
-client.commands = new Discord.Collection();
-client.aliases = new Discord.Collection();
+client.commands = new Collection();
+client.aliases = new Collection();
 
 /**
  * init()