+++ mongo.js
@@ -5,8 +5,7 @@ module.exports = async () => {
     await mongoose.connect(mongoPath, {
         keepAlive: true,
         useNewUrlParser: true,
-        useUnifiedTopology: true,
-        useFindAndModify: false,
+        useUnifiedTopology: true
     });
 
     return mongoose;